requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.9.0=pypi_0
aiohttp=3.6.2=pypi_0
alembic=1.4.2=pypi_0
aniso8601=8.0.0=pypi_0
astor=0.8.1=pypi_0
async-timeout=3.0.1=pypi_0
attrs=19.3.0=pypi_0
backcall=0.1.0=pypi_0
beautifulsoup4=4.9.1=pypi_0
bert-embedding=1.0.1=pypi_0
bert-for-tf2=0.14.4=pypi_0
bert-tensorflow=1.0.1=pypi_0
blas=1.0=mkl
bleach=3.1.1=pypi_0
blis=0.4.1=pypi_0
boto=2.49.0=py36_0
boto3=1.12.27=pypi_0
botocore=1.15.27=pypi_0
ca-certificates=2020.1.1=0
cachetools=4.0.0=pypi_0
cairocffi=1.1.0=pypi_0
catalogue=1.0.0=pypi_0
catboost=0.24.1=pypi_0
certifi=2020.4.5.1=py36_0
cffi=1.14.0=py36h2e261b9_0
chardet=3.0.4=py36_1003
click=7.1.2=pypi_0
cloudpickle=1.5.0=pypi_0
colorama=0.4.3=pypi_0
colorlog=4.1.0=pypi_0
cryptography=2.9.2=py36h1ba5d50_0
cycler=0.10.0=pypi_0
cymem=2.0.3=pypi_0
databricks-cli=0.11.0=pypi_0
dataclasses=0.7=pypi_0
decorator=4.4.2=pypi_0
defusedxml=0.6.0=pypi_0
dill=0.3.2=pypi_0
docker=4.2.2=pypi_0
docutils=0.15.2=pypi_0
dotmap=1.3.0=pypi_0
email=6.0.0a1=pypi_0
emoji=0.5.4=pypi_0
emoji-extractor=1.0.17=pypi_0
emojis=0.5.1=pypi_0
entrypoints=0.3=pypi_0
farm=0.4.3=pypi_0
filelock=3.0.12=pypi_0
flask=1.1.2=pypi_0
flask-cors=3.0.8=pypi_0
flask-restplus=0.13.0=pypi_0
funcy=1.14=pypi_0
future=0.18.2=pypi_0
gab=0.0.1=pypi_0
gast=0.2.2=pypi_0
gensim=3.8.3=pypi_0
gitdb=4.0.5=pypi_0
gitpython=3.1.7=pypi_0
gluonnlp=0.6.0=pypi_0
google=2.0.3=pypi_0
google-api-core=1.16.0=pypi_0
google-auth=1.11.3=pypi_0
google-auth-oauthlib=0.4.1=pypi_0
google-cloud-core=1.3.0=py_0
google-cloud-storage=1.26.0=pypi_0
google-pasta=0.2.0=pypi_0
google-resumable-media=0.5.0=py_1
googleapis-common-protos=1.51.0=py36_2
grpcio=1.28.1=pypi_0
gunicorn=20.0.4=pypi_0
h5py=2.10.0=pypi_0
idna=2.9=py_1
idna-ssl=1.1.0=pypi_0
importlib-metadata=1.5.0=pypi_0
intel-openmp=2020.1=217
ipykernel=5.1.4=pypi_0
ipython=7.13.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.5.1=pypi_0
itsdangerous=1.1.0=pypi_0
jedi=0.16.0=pypi_0
jinja2=2.11.1=pypi_0
jmespath=0.9.5=pypi_0
joblib=0.14.1=pypi_0
jsonschema=3.2.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=6.0.0=pypi_0
jupyter-console=6.1.0=pypi_0
jupyter-contrib-core=0.3.3=pypi_0
jupyter-contrib-nbextensions=0.5.1=pypi_0
jupyter-core=4.6.3=pypi_0
jupyter-highlight-selected-word=0.2.0=pypi_0
jupyter-latex-envs=1.4.6=pypi_0
jupyter-nbextensions-configurator=0.4.1=pypi_0
keras=2.3.1=pypi_0
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.0=pypi_0
kiwisolver=1.1.0=pypi_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libprotobuf=3.11.4=hd408876_0
libstdcxx-ng=9.1.0=hdf63c60_0
lightgbm=3.0.0=pypi_0
lxml=4.5.0=pypi_0
mako=1.1.3=pypi_0
markdown=3.2.1=pypi_0
markupsafe=1.1.1=pypi_0
matplotlib=3.2.1=pypi_0
mistune=0.8.4=pypi_0
mittens=0.2=pypi_0
mkl=2020.1=217
mkl-service=2.3.0=py36he904b0f_0
mkl_fft=1.0.15=py36ha843d7b_0
mkl_random=1.1.0=py36hd6b4f25_0
mlflow=1.0.0=pypi_0
more-itertools=8.2.0=pypi_0
msgpack=0.6.2=pypi_0
multidict=4.7.6=pypi_0
murmurhash=1.0.2=pypi_0
mxnet=1.4.0=pypi_0
nbconvert=5.6.1=pypi_0
nbformat=5.0.4=pypi_0
ncurses=6.2=he6710b0_0
networkx=2.4=pypi_0
nltk=3.4.5=pypi_0
notebook=6.0.3=pypi_0
numexpr=2.7.1=pypi_0
numpy=1.16.1=pypi_0
oauthlib=3.1.0=pypi_0
onnx=1.7.0=pypi_0
onnxruntime=1.3.0=pypi_0
opencv-python=4.2.0.34=pypi_0
openssl=1.1.1g=h7b6447c_0
opt-einsum=3.2.0=pypi_0
packaging=20.3=pypi_0
pandas=1.0.3=pypi_0
pandocfilters=1.4.2=pypi_0
params-flow=0.8.2=pypi_0
parso=0.6.2=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=7.2.0=pypi_0
pip=20.0.2=py36_1
plac=1.1.3=pypi_0
plotly=4.9.0=pypi_0
pluggy=0.13.1=pypi_0
preshed=3.0.2=pypi_0
prometheus-client=0.7.1=pypi_0
prompt-toolkit=3.0.4=pypi_0
protobuf=3.11.3=pypi_0
ptyprocess=0.6.0=pypi_0
py=1.8.1=pypi_0
py-params=0.9.7=pypi_0
py-spy=0.3.3=pypi_0
pyarrow=1.0.1=pypi_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.20=py_0
pydot=1.4.1=pypi_0
pygments=2.6.1=pypi_0
pyldavis=2.1.2=pypi_0
pyopenssl=19.1.0=py36_0
pyparsing=2.4.6=pypi_0
pyphen=0.9.5=pypi_0
pyrsistent=0.15.7=pypi_0
pysocks=1.7.1=py36_0
pytest=5.4.1=pypi_0
python=3.6.10=h0371630_0
python-dateutil=2.8.1=py_0
python-editor=1.0.4=pypi_0
python-graphviz=0.8.4=pypi_0
python-igraph=0.7.1.post6=pypi_0
pytorch-pretrained-bert=0.6.2=pypi_0
pytrends=4.7.2=pypi_0
pytz=2019.3=pypi_0
pyyaml=5.3.1=pypi_0
pyzmq=19.0.0=pypi_0
qtconsole=4.7.1=pypi_0
qtpy=1.9.0=pypi_0
querystring-parser=1.2.4=pypi_0
rauth=0.7.3=pypi_0
ray=0.8.5=pypi_0
readline=7.0=h7b6447c_5
redis=3.4.1=pypi_0
regex=2020.4.4=pypi_0
requests=2.23.0=py36_0
requests-oauthlib=1.3.0=pypi_0
retrying=1.3.3=pypi_0
rsa=4.0=py_0
s3transfer=0.3.3=pypi_0
sacremoses=0.0.43=pypi_0
scikit-learn=0.22.2.post1=pypi_0
scipy=1.1.0=pypi_0
seaborn=0.11.0=pypi_0
send2trash=1.5.0=pypi_0
sentencepiece=0.1.91=pypi_0
seqeval=0.0.12=pypi_0
setuptools=46.3.1=pypi_0
simplejson=3.17.0=pypi_0
six=1.14.0=py36_0
sklearn=0.0=pypi_0
slackclient=2.5.0=pypi_0
smart-open=1.10.0=pypi_0
smart_open=2.0.0=py_0
smmap=3.0.4=pypi_0
soupsieve=2.0.1=pypi_0
spacy=2.2.4=pypi_0
sqlalchemy=1.3.18=pypi_0
sqlite=3.31.1=h7b6447c_0
sqlparse=0.3.1=pypi_0
srsly=1.0.2=pypi_0
tabulate=0.8.7=pypi_0
tb-nightly=2.3.0a20200717=pypi_0
tensorboard=2.1.1=pypi_0
tensorflow=2.1.0=pypi_0
tensorflow-estimator=2.1.0=pypi_0
tensorflow-hub=0.8.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.8.3=pypi_0
testpath=0.4.4=pypi_0
textstat=0.6.2=pypi_0
texttable=1.6.2=pypi_0
tf-estimator-nightly=2.4.0.dev2020071801=pypi_0
tf-nightly-gpu=2.4.0.dev20200718=pypi_0
thinc=7.4.0=pypi_0
tk=8.6.8=hbc83047_0
tokenizers=0.5.2=pypi_0
torch=1.4.0=pypi_0
tornado=6.0.4=pypi_0
tqdm=4.43.0=pypi_0
traitlets=4.3.3=pypi_0
transformers=2.7.0=pypi_0
tweepy=3.8.0=pypi_0
typing=3.6.6=pypi_0
typing-extensions=3.7.4.2=pypi_0
urllib3=1.25.8=py36_0
vadersentiment=3.3.2=pypi_0
wasabi=0.6.0=pypi_0
wcwidth=0.1.8=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=0.57.0=pypi_0
werkzeug=0.16.1=pypi_0
wheel=0.34.2=py36_0
whoosh=2.7.4=pypi_0
widgetsnbextension=3.5.1=pypi_0
wordcloud=1.8.0=pypi_0
wrapt=1.12.1=pypi_0
xgboost=1.2.0=pypi_0
xz=5.2.4=h14c3975_4
yarl=1.4.2=pypi_0
zipp=3.1.0=pypi_0
zlib=1.2.11=h7b6447c_3
